座標取得処理をlibraryに移動
