+++ plugins/hato.py
@@ -3,14 +3,15 @@
 """hatobotのチャット部分"""
 
 import imghdr
-import json
 import os
 import re
 from logging import getLogger
 from tempfile import NamedTemporaryFile
-from typing import List, Dict, Optional
+from typing import List
+
 import requests
 import slackbot_settings as conf
+from library.amesh import get_geo_data
 from library.vocabularydb import get_vocabularys, add_vocabulary, show_vocabulary, delete_vocabulary, show_random_vocabulary
 from library.earthquake import generate_quake_info_for_slack, get_quake_list
 from library.hukidasi import generator
@@ -134,31 +135,6 @@ def weather_map_url(appid: str, lat: str, lon: str) -> str:
         'https://map.yahooapis.jp/map/V1/static?' +
         'appid={}&lat={}&lon={}&z=12&height=640&width=800&overlay=type:rainfall|datelabel:off'
     ).format(appid, lat, lon)
-
-
-def get_geo_data(place: str) -> Optional[Dict[str, str]]:
-    """
-    地名や住所から座標を取得する
-    :param place: 地名・住所
-    :return: place: 地名, lat: 緯度, lon: 経度
-    """
-    geo_data = json.loads(requests.get('https://map.yahooapis.jp/geocode/V1/geoCoder',
-                                       {
-                                           'appid': conf.YAHOO_API_TOKEN,
-                                           'query': place,
-                                           'output': 'json'
-                                       }).content)
-
-    if 'Feature' in geo_data:
-        for feature in geo_data['Feature']:
-            if 'Name' in feature and feature['Name']:
-                if 'Geometry' in feature and feature['Geometry']:
-                    geometry = feature['Geometry']
-                    if 'Coordinates' in geometry and geometry['Coordinates']:
-                        lon, lat = geometry['Coordinates'].split(',', maxsplit=2)
-                        return {'place': feature['Name'], 'lat': lat, 'lon': lon}
-
-    return None
 
 
 def amesh(place: str):
